requirements.txt
flask
werkzeug
python-dotenv
langchain-google-genai
transformers
moviepy
gTTS
pydub
openai